s390x: remove the volume
When the runner tries to auto-update, it may enter and endless loop when
a volume stanza is set to workdir.

Based on https://docs.docker.com/engine/reference/builder/#volume

> The docker run command initializes the newly created volume with any data that exists at the specified location within the base image.

So, when we have a container with version X, the runner may start, see
it needs version X+1, upgrade and restarts.
In ephemeral mode, this results in a containers restart, remounting the
volumes with a copy of the content in the image, e.g version X, which
needs update.....